% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vegetables.srilanka.R
\docType{data}
\name{vegetables.srilanka}
\alias{vegetables.srilanka}
\title{Daily wholesale and retail vegetable prices in Sri Lanka}
\format{
A tibble with 62908 rows and 5 variables:
\describe{
  \item{Date}{Date}
  \item{Item}{Vegetable name}
  \item{Type}{Wholesale or Retail price}
  \item{Market}{Pettah or Dambulla market}
  \item{Price}{Price in LKR per kg}
  }
}
\source{
Accessed from Daily Reports - Central Bank of Sri Lanka
}
\usage{
vegetables.srilanka
}
\description{
Daily wholesale and retail vegetable prices at Dambulla and Petta markets in Sri Lanka
}
\examples{
head(vegetables.srilanka)
}
\keyword{datasets}
